requirements-updater.txt
PyMongo==3.7.1
pyyaml==3.13
#-e git+https://github.com/BenjiBackslash/pika_client#egg=
pika_client==0.9
